feat: 검색창 디자인 변경 및 기능 추가 (close #64)
- 검색시 검색바 근처에서 검색결과 출력
- 검색결과 컨테이너가 출력되고 있는 상태에서 외부 누르면 컨테이너
  안보이게 처리
 - 컨테이너 크기에 따라 상위 8개만 검색결과 출력
