--- src/components/common/search/SearchData.jsx
@@ -1,1 +1,0 @@
-const SearchData = () => {};
