--- src/components/search/Search.jsx
@@ -1,11 +1,0 @@
-
-
-const Search = () => {
-    return (
-        <div>
-            
-        </div>
-    );
-};
-
-export default Search;
